Creates a file for dynamic memory managment
All allocate/deallocate and initialization of arrays are moved in
a new file,tdymemory.c
